Move code as-is to webpages directory
