+++ background/background.js
@@ -30,4 +30,3 @@ import "./get-popups.js";
 // Respond to requests by the settings page and others
 import "./handle-settings-page.js";
 import "./handle-licenses.js";
-import "./handle-unsupported-version.js";
